invalid scenarios using scenario outline
